+++ pkg/app/app.go
@@ -16,7 +16,7 @@ type Page struct {
 	//每页数量
 	PageSize int `json:"page_size"`
 	//总行数
-	TotalRows int `json:"total_rows"`
+	TotalRows int64 `json:"total_rows"`
 }
 
 type SuccessList struct {
@@ -58,7 +58,7 @@ func (r *Response) ToResponse(data interface{}) {
 }
 
 //响应多条分页json
-func (r *Response) ToResponseList(list interface{}, totalRows int) {
+func (r *Response) ToResponseList(list interface{}, totalRows int64) {
 	success := errcode.Success
 	response := gin.H{
 		"code": success.Code(),
